+++ CHANGELOG.md
@@ -11,6 +11,7 @@ All notable changes to this project will be documented in this file.
 - Renamed `rdtomo.transformers` to `rdtomo.coords`
 - `rdtomo` functions should now use the `rdtomo.coords.Pos` class in most cases, except for internal calculations before initiating a `Pos` object (**exception**: `rdtomo.trackfinding` which requires a more thorough walkthrough)
 - Changed the name of `rdtomo.gnss.rtkp` to `rdtomo.gnss.ppk` to better match reality
+- `rdtomo.ppk` and `rdtomo.ppp` now return a tuple with the `Pos` object as the first entry and a dict with the rest of the results as the second
 
 ## [0.1.0] - 2025-11-28